Backend/requirements.txt
sqlalchemy==2.0.35
aiosqlite==0.20.0
cryptography==43.0.1
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
PyMuPDF==1.24.10
Pillow==10.4.0
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
python-dotenv==1.0.1
slowapi==0.1.9
loguru==0.7.2
langdetect==1.0.9
opencv-python>=4.8.0
numpy>=1.26.0
pytesseract>=0.3.10
faster-whisper>=1.0.0
googlemaps>=4.10.0
anthropic>=0.30.0
reportlab>=4.2.0
polyline>=2.0.0
